moving image formats into submodules (for Point Inside)
