+++ src/host-modules/aku_modules_config.h
@@ -33,6 +33,22 @@
 		#define AKU_WITH_HTTP_SERVER 1
 	#endif
 
+	#ifndef AKU_WITH_IMAGE_JPG
+		#define AKU_WITH_IMAGE_JPG 1
+	#endif
+
+	#ifndef AKU_WITH_IMAGE_PNG
+		#define AKU_WITH_IMAGE_PNG 1
+	#endif
+
+	#ifndef AKU_WITH_IMAGE_PVR
+		#define AKU_WITH_IMAGE_PVR 1
+	#endif
+
+	#ifndef AKU_WITH_IMAGE_WEBP
+		#define AKU_WITH_IMAGE_WEBP 1
+	#endif
+
 	#ifndef AKU_WITH_LUAEXT
 		#define AKU_WITH_LUAEXT 1
 	#endif
