+++ src/moai-image-webp/host.h
@@ -1,0 +1,14 @@
+//----------------------------------------------------------------//
+// Copyright (c) 2010-2011 Zipline Games, Inc. 
+// All Rights Reserved. 
+// http://getmoai.com
+//----------------------------------------------------------------//
+
+#ifndef MOAI_IMAGE_WEBP_HOST_H
+#define MOAI_IMAGE_WEBP_HOST_H
+
+#include <moai-core/host.h>
+
+AKU_API void	AKUImageWebPContextInitialize		();
+
+#endif
